requirements.txt
# Add required packages here
pandas
numpy
scikit-learn
matplotlib
seaborn
pyyaml